implement calculate data into calculator fragment

+++ app/src/main/java/com/example/foodnote/data/interactor/diary_item_detail_interactor/DiaryItemDetailInteractorImpl.kt
@@ -31,19 +31,19 @@ class DiaryItemDetailInteractorImpl(private val repo: DiaryItemDetailRepository)
         return withContext(Dispatchers.Default) {
             val calorieSum = list.sumOf {
                 it.kiloCalories * list.size
-            }
+            }.toInt()
             val proteinSum = list.sumOf {
                 it.protein * it.count
-            }
+            }.toInt()
 
             val fatSum = list.sumOf {
                 it.fat * it.count
-            }
+            }.toInt()
 
             val carbSum = list.sumOf {
                 it.carbohydrate * it.count
-            }
-            TotalFoodResult(0.0, calorieSum, 0.0, proteinSum, 0.0, fatSum, 0.0, carbSum)
+            }.toInt()
+            TotalFoodResult(0, calorieSum, 0, proteinSum, 0, fatSum, 0, carbSum)
         }
     }
